algorithm simplify
