+++ .vscode/settings.json
@@ -22,6 +22,7 @@
     "python.analysis.typeCheckingMode": "basic",
     "C_Cpp.formatting": "default",
     "editor.defaultFormatter": "esbenp.prettier-vscode",
+    "prettier.tabWidth": 4,
     "[json]": {
         "editor.defaultFormatter": "vscode.json-language-features"
     },
