+++ src/bw_bringup/launch/playback.launch
@@ -3,7 +3,7 @@
     <arg name="robot" default="$(env ROBOT)" />
     <arg name="model"
         default="$(find bw_data)/data/models/nhrl_dataset_mask_rcnn_R_50_FPN_3x_2023-11-20" />
-    <arg name="svo_name" default="mini_bot_2024-02-17T16-14-33.svo" />
+    <arg name="svo_name" default="mini_bot_2024-02-24T12-56-11.svo" />
 
     <include file="$(find bw_bringup)/launch/subsystems/common.launch">
         <arg name="auto_initialize" value="true" />
