+++ src/bw_navigation/config/costmap_common_params.yaml
@@ -1,7 +1,7 @@
 ### Common costmap parameters ###
 global_frame: map
 robot_base_frame: base_link_mini_bot
-resolution: 0.05
+resolution: 0.025
 rolling_window: false
 update_frequency: 3.0
 publish_frequency: 1.0
