+++ src/bw_navigation/src/bw_navigation/selector_algorithms/__init__.py
@@ -1,3 +1,4 @@
+from .crash_avoid_front_selector import CrashAvoidFrontSelector
 from .crash_selector import CrashSelector
 from .push_from_behind_selector import PushFromBehindSelector
 from .sacrificial_selector import SacrificialSelector
