+++ src/bw_navigation/src/bw_navigation/target_selector.py
@@ -14,7 +14,12 @@ from sensor_msgs import point_cloud2
 from sensor_msgs.msg import PointCloud2, PointField
 from std_msgs.msg import Header
 
-from bw_navigation.selector_algorithms import CrashSelector, PushFromBehindSelector, SacrificialSelector
+from bw_navigation.selector_algorithms import (
+    CrashAvoidFrontSelector,
+    CrashSelector,
+    PushFromBehindSelector,
+    SacrificialSelector,
+)
 from bw_navigation.selector_algorithms.base_selector import BaseSelector
 from bw_navigation.selector_algorithms.match_state import MatchState
 
@@ -39,6 +44,7 @@ class TargetSelector:
             "sacrificial_selector": SacrificialSelector,
             "push_from_behind_selector": PushFromBehindSelector,
             "crash_selector": CrashSelector,
+            "crash_avoid_front_selector": CrashAvoidFrontSelector,
         }[self.algorithm_name]()
 
         all_robots = RobotFleetConfig.from_dict(robot_config)
